clienthub/requirements.txt
BeautifulSoup==3.2.1
Crowd==0.3dev
FormEncode==1.2.4
Jinja2==2.6
PyYAML==3.10
argparse==1.2.1
beatbox==20.0
-e git+git@github.com:10gen/python-libs.git@93819b45847349abab00366830e7a2b5972d59d3#egg=corplibs-dev
ipython==0.13
jira-python==0.12
markdown2==2.1.0
pymongo==2.3
python-magic==0.4.2
pytz==2012f
requests==0.14.1
suds==0.4
tlslite==0.4.1
web.py==0.37
wsgiref==0.1.2

